Deleting a comment: service method
